requirements.txt
confluent_kafka==1.9.2
influx_line_protocol==0.1.5
influxdb_client==1.34.0
PyYAML==6.0
requests==2.28.1
schema==0.7.5
shapes==0.1.3
certifi

